refactor: extracted AccountData component
